pyproject.toml
[project]
name = "tavily_web_eval_generator"
version = "0.0.1"
description = "Template for a web search rag dataset generator using Tavily and Langgraph."
authors = [
    { name = "Tavily", email = "support@tavily.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "langchain-core",
    "langchain-openai",
    "langgraph",
    "tavily-python",
    "pydantic",
    "langsmith"
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tavily.templates.web_eval_generator", "web_eval_generator"]
[tool.setuptools.package-dir]
"tavily.templates.web_eval_generator" = "src/web_eval_generator"
"web_eval_generator" = "src/web_eval_generator"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007", # We actually do want to import from typing_extensions
    "UP035", # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"